requirements.txt
python-telegram-bot[job-queue]>=20.0
pytz>=2023.3
python-dotenv>=1.0.0
